requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6 ; platform_system == "Windows"
discord==2.3.2
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
openai>=1.0.0
requests==2.31.0
tqdm==4.65.0
urllib3==2.0.4
yarl==1.9.2
faiss-cpu
sentence-transformers
dotenv
